# api/openapi.yaml
openapi: 3.0.0
info:
  description: LocationIQ provides flexible enterprise-grade location based solutions. We work with developers, startups and enterprises worldwide serving billions of requests everyday. This page provides an overview of the technical aspects of our API and will help you get started.
  title: LocationIQ
  version: 1.0.1
servers:
- url: https://eu1.locationiq.com/v1
paths:
  /search.php:
    get:
      description: The Search API allows converting addresses, such as a street address, into geographic coordinates (latitude and longitude). These coordinates can serve various use-cases, from placing markers on a map to helping algorithms determine nearby bus stops. This process is also known as Forward Geocoding.
      operationId: search
      parameters:
      - description: Address to geocode
        example: '"Empire state building"'
        explode: true
        in: query
        name: q
        required: true
        schema:
          type: string
        style: form
      - description: Format to geocode. Only JSON supported for SDKs
        example: '"json"'
        explode: true
        in: query
        name: format
        required: true
        schema:
          enum:
          - json
          type: string
        style: form
      - description: For responses with no city value in the address section, the next available element in this order - city_district, locality, town, borough, municipality, village, hamlet, quarter, neighbourhood - from the address section will be normalized to city. Defaults to 1 for SDKs.
        example: 1
        explode: true
        in: query
        name: normalizecity
        required: true
        schema:
          enum:
          - 1
          format: int32
          type: integer
        style: form
      - description: Include a breakdown of the address into elements. Defaults to 0.
        example: 1
        explode: true
        in: query
        name: addressdetails
        required: false
        schema:
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      - description: 'The preferred area to find search results.  To restrict results
          to those within the viewbox, use along with the bounded option. Tuple of
          4 floats. Any two corner points of the box - `max_lon,max_lat,min_lon,min_lat`
          or `min_lon,min_lat,max_lon,max_lat` - are accepted in any order as long
          as they span a real box. '
        example: '"-132.84908,47.69382,-70.44674,30.82531"'
        explode: true
        in: query
        name: viewbox
        required: false
        schema:
          type: string
        style: form
      - description: Restrict the results to only items contained with the viewbox
        example: 1
        explode: true
        in: query
        name: bounded
        required: false
        schema:
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      - description: Limit the number of returned results. Default is 10.
        example: 10
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          format: int32
          type: integer
        style: form
      - description: Preferred language order for showing search results, overrides the value specified in the Accept-Language HTTP header. Defaults to en. To use native language for the response when available, use accept-language=native
        example: '"en"'
        explode: true
        in: query
        name: accept-language
        required: false
        schema:
          type: string
        style: form
      - description: Limit search to a list of countries.
        example: '"us"'
        explode: true
        in: query
        name: countrycodes
        required: false
        schema:
          type: string
        style: form
      - description: Include a list of alternative names in the results. These may include language variants, references, operator and brand.
        example: 1
        explode: true
        in: query
        name: namedetails
        required: false
        schema:
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      - description: Sometimes you have several objects in OSM identifying the same place or object in reality. The simplest case is a street being split in many different OSM ways due to different characteristics. Nominatim will attempt to detect such duplicates and only return one match; this is controlled by the dedupe parameter which defaults to 1. Since the limit is, for reasons of efficiency, enforced before and not after de-duplicating, it is possible that de-duplicating leaves you with less results than requested.
        example: 1
        explode: true
        in: query
        name: dedupe
        required: false
        schema:
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      - description: Include additional information in the result if available, e.g. wikipedia link, opening hours.
        example: 0
        explode: true
        in: query
        name: extratags
        required: false
        schema:
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      - description: Adds state or province code when available to the statecode key inside the address element. Currently supported for addresses in the USA, Canada and Australia. Defaults to 0
        example: 0
        explode: true
        in: query
        name: statecode
        required: false
        schema:
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                  - place_id: "83874904"
                    licence: © LocationIQ.com CC BY 4.0, Data © OpenStreetMap contributors, ODbL 1.0
                    osm_type: way
                    osm_id: "34633854"
                    boundingbox:
                    - "40.7479226"
                    - "40.7489422"
                    - -73.9864855
                    - -73.9848259
                    lat: "40.7484284"
                    lon: -73.9856546198733
                    display_name: Empire State Building, 350, 5th Avenue, Korea Town, Midtown South, Manhattan, Manhattan Community Board 5, New York County, New York City, New York, 10001, United States of America
                    class: tourism
                    type: attraction
                    importance: 0.43072416348613
                    icon: https://locationiq.org/static/images/mapicons/poi_point_of_interest.p.20.png
                    address:
                      name: Empire State Building
                      house_number: "350"
                      road: 5th Avenue
                      commercial: Korea Town
                      suburb: Midtown South
                      city_district: Manhattan
                      city: New York City
                      county: New York County
                      state: New York
                      postcode: "10001"
                      country: United States of America
                      country_code: us
                  - place_id: "28210682"
                    licence: © LocationIQ.com CC BY 4.0, Data © OpenStreetMap contributors, ODbL 1.0
                    osm_type: node
                    osm_id: "2709306673"
                    boundingbox:
                    - "40.7481128"
                    - "40.7482128"
                    - -73.9850458
                    - -73.9849458
                    lat: "40.7481628"
                    lon: -73.9849958
                    display_name: Empire State Building, 350, 5th Avenue, Korea Town, Midtown South, Manhattan, Manhattan Community Board 5, New York County, New York City, New York, 10001, United States of America
                    class: tourism
                    type: attraction
                    importance: 0.30025
                    icon: https://locationiq.org/static/images/mapicons/poi_point_of_interest.p.20.png
                    address:
                      name: Empire State Building
                      house_number: "350"
                      road: 5th Avenue
                      commercial: Korea Town
                      suburb: Midtown South
                      city_district: Manhattan
                      city: New York City
                      county: New York County
                      state: New York
                      postcode: "10001"
                      country: United States of America
                      country_code: us
                  - place_id: "66500562"
                    licence: © LocationIQ.com CC BY 4.0, Data © OpenStreetMap contributors, ODbL 1.0
                    osm_type: node
                    osm_id: "5552887205"
                    boundingbox:
                    - "52.2956835"
                    - "52.2957835"
                    - "16.7553074"
                    - "16.7554074"
                    lat: "52.2957335"
                    lon: "16.7553574"
                    display_name: Empire State Building, Poznańska, Trzebaw, gmina Stęszew, powiat poznański, Greater Poland Voivodeship, Poland
                    class: tourism
                    type: artwork
                    importance: 0.30025
                    icon: https://locationiq.org/static/images/mapicons/tourist_art_gallery2.p.20.png
                    address:
                      name: Empire State Building
                      road: Poznańska
                      village: Trzebaw
                      county: gmina Stęszew
                      state: Greater Poland Voivodeship
                      country: Poland
                      country_code: pl
                      city: Trzebaw
              schema:
                items:
                  $ref: '#/components/schemas/location'
                title: fowardresponse
                type: array
          description: OK
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Invalid Request
              schema:
                $ref: '#/components/schemas/error'
          description: Required parameters are missing, or invalid
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Invalid key
              schema:
                $ref: '#/components/schemas/error'
          description: An invalid API key was provided / Key not active
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Unable to geocode
              schema:
                $ref: '#/components/schemas/error'
          description: No location or places were found for the given input
        429:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Rate Limited Second
              schema:
                $ref: '#/components/schemas/error'
          description: Request exceeded the rate-limits set on your account
        500:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Unknown error - Please try again after some time
              schema:
                $ref: '#/components/schemas/error'
          description: This is an error on the server's side, we monitor this 24x7 and you should try again.
      security:
      - key: []
      summary: Forward Geocoding
      tags:
      - Search
      x-swagger-router-controller: Search
  /reverse.php:
    get:
      description: Reverse geocoding is the process of converting a coordinate or location (latitude, longitude) to a readable address or place name. This permits the identification of nearby street addresses, places, and/or area subdivisions such as a neighborhood, county, state, or country.
      operationId: reverse
      parameters:
      - description: Latitude of the location to generate an address for.
        example: 17.24
        explode: true
        in: query
        name: lat
        required: true
        schema:
          maximum: 90
          minimum: -90
          type: number
        style: form
      - description: Longitude of the location to generate an address for.
        example: 74.25
        explode: true
        in: query
        name: lon
        required: true
        schema:
          maximum: 180
          minimum: -180
          type: number
        style: form
      - description: Format to geocode. Only JSON supported for SDKs
        example: '"json"'
        explode: true
        in: query
        name: format
        required: true
        schema:
          enum:
          - json
          type: string
        style: form
      - description: Normalizes village to city level data to city
        example: 1
        explode: true
        in: query
        name: normalizecity
        required: true
        schema:
          enum:
          - 1
          format: int32
          type: integer
        style: form
      - description: Include a breakdown of the address into elements. Defaults to 1.
        example: 1
        explode: true
        in: query
        name: addressdetails
        required: false
        schema:
          default: 1
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      - description: Preferred language order for showing search results, overrides the value specified in the Accept-Language HTTP header. Defaults to en. To use native language for the response when available, use accept-language=native
        example: '"en"'
        explode: true
        in: query
        name: accept-language
        required: false
        schema:
          type: string
        style: form
      - description: Include a list of alternative names in the results. These may include language variants, references, operator and brand.
        example: 0
        explode: true
        in: query
        name: namedetails
        required: false
        schema:
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      - description: Include additional information in the result if available, e.g. wikipedia link, opening hours.
        example: 0
        explode: true
        in: query
        name: extratags
        required: false
        schema:
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      - description: Adds state or province code when available to the statecode key inside the address element. Currently supported for addresses in the USA, Canada and Australia. Defaults to 0
        example: 0
        explode: true
        in: query
        name: statecode
        required: false
        schema:
          enum:
          - 0
          - 1
          format: int32
          type: integer
        style: form
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                    place_id: "83874904"
                    licence: © LocationIQ.com CC BY 4.0, Data © OpenStreetMap contributors, ODbL 1.0
                    osm_type: way
                    osm_id: "34633854"
                    lat: "40.7484284"
                    lon: -73.9856546198733
                    display_name: Empire State Building, 350, 5th Avenue, Korea Town, Midtown South, Manhattan, Manhattan Community Board 5, New York County, New York City, New York, 10001, United States of America
                    address:
                      name: Empire State Building
                      house_number: "350"
                      road: 5th Avenue
                      commercial: Korea Town
                      suburb: Midtown South
                      city_district: Manhattan
                      city: New York City
                      county: New York County
                      state: New York
                      postcode: "10001"
                      country: United States of America
                      country_code: us
                    boundingbox:
                    - "40.7479226"
                    - "40.7489422"
                    - -73.9864855
                    - -73.9848259
              schema:
                $ref: '#/components/schemas/location'
          description: OK
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Invalid Request
              schema:
                $ref: '#/components/schemas/error'
          description: Required parameters are missing, or invalid
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Key not active - Please write to hello@locationiq.org
              schema:
                $ref: '#/components/schemas/error'
          description: An invalid API key was provided / Key not active
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Unable to geocode
              schema:
                $ref: '#/components/schemas/error'
          description: No location or places were found for the given input
        429:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Rate Limited Second
              schema:
                $ref: '#/components/schemas/error'
          description: Too Many Requests
        500:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Unknown error - Please try again after some time
              schema:
                $ref: '#/components/schemas/error'
          description: Unknown error - Please try again after some time
      security:
      - key: []
      summary: Reverse Geocoding
      tags:
      - reverse
      x-swagger-router-controller: Reverse
  /balance.php:
    get:
      description: The Balance API provides a count of request credits left in the user's account for the day. Balance is reset at midnight UTC everyday (00:00 UTC).
      operationId: balance
      responses:
        200:
          content:
            application/json:
              examples:
                response:
                  value:
                    status: ok
                    balance:
                      day: 30000
              schema:
                $ref: '#/components/schemas/balance'
          description: OK
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Invalid Request
              schema:
                $ref: '#/components/schemas/error'
          description: Bad Request
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Invalid key
              schema:
                $ref: '#/components/schemas/error'
          description: Unauthorized
        429:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Rate Limited Day
              schema:
                $ref: '#/components/schemas/error'
          description: Too Many Requests
        500:
          content:
            application/json:
              examples:
                response:
                  value:
                    error: Unknown error - Please try again after some time
              schema:
                $ref: '#/components/schemas/error'
          description: Internal Server Error
      security:
      - key: []
      tags:
      - Balance
      x-swagger-router-controller: Balance
components:
  schemas:
    namedetails:
      example:
        name: Empire State Building
      properties:
        name:
          type: string
      title: namedetails
      type: object
    address:
      example:
        house_number: "3894"
        road: Spring Mill Way
        residential: Hunter’s Point
        village: Landen
        city: Landen
        county: Warren County
        state: Ohio
        postcode: "45039"
        country: United States of America
        country_code: us
        state_code: oh
      properties:
        house_number:
          type: string
        road:
          type: string
        residential:
          type: string
        borough:
          type: string
        neighbourhood:
          type: string
        quarter:
          type: string
        hamlet:
          type: string
        suburb:
          type: string
        island:
          type: string
        village:
          type: string
        town:
          type: string
        city:
          type: string
        city_district:
          type: string
        county:
          type: string
        state:
          type: string
        state_district:
          type: string
        postcode:
          type: string
        country:
          type: string
        country_code:
          type: string
        state_code:
          type: string
      title: address
      type: object
    balance:
      example:
        status: ok
        balance:
          day: 547933
          bonus: 123
      properties:
        status:
          type: string
        balance:
          $ref: '#/components/schemas/daybalance'
      title: balance
      type: object
    daybalance:
      properties:
        day:
          format: int32
          type: integer
        bonus:
          format: int32
          type: integer
      title: daybalance
      type: object
    error:
      example:
        error: Invalid key
      properties:
        error:
          type: string
      title: error
      type: object
    location:
      example:
        place_id: "223483692"
        licence: © LocationIQ.com CC BY 4.0, Data © OpenStreetMap contributors, ODbL 1.0
        osm_type: way
        osm_id: "19301621"
        boundingbox:
        - "39.307405567782"
        - "39.307505567782"
        - -84.292824851595
        - -84.292724851595
        lat: "39.3074555677816"
        lon: -84.2927748515948
        display_name: 3894, Spring Mill Way, Hunter’s Point, Landen, Warren County, Ohio, 45039, United States of America
        class: place
        type: house
        importance: 0.62025
        address:
          house_number: "3894"
          road: Spring Mill Way
          residential: Hunter’s Point
          village: Landen
          county: Warren County
          state: Ohio
          postcode: "45039"
          country: United States of America
          country_code: us
          city: Landen
      properties:
        place_id:
          type: string
        licence:
          type: string
        osm_type:
          type: string
        osm_id:
          type: string
        boundingbox:
          items:
            type: string
          type: array
        lat:
          type: string
        lon:
          type: string
        display_name:
          type: string
        class:
          type: string
        type:
          type: string
        importance:
          type: number
        address:
          $ref: '#/components/schemas/address'
        namedetails:
          $ref: '#/components/schemas/namedetails'
      title: location
      type: object
  securitySchemes:
    key:
      in: query
      name: key
      type: apiKey